+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog of lizard-nxt client
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
-- Fix IE svg not scalin well, set height in css.
+- Fix IE svg not scaling well, set height in css.
+
+- Fix IE dashoard dragging, inlude datalist polyfill and get attribute instead of element.dataset.
 
 
 Release 3.0.10 (2016-4-22)