+++ CHANGES.rst
@@ -5,6 +5,8 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 - Fix respect to layer minZoom and maxZoom.
 
+- Login button visible if not authenticated.
+
 
 Release 3.0.10 (2016-4-22)
 ---------------------