+++ CHANGES.rst
@@ -4,9 +4,12 @@ Changelog of lizard-nxt client
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
+- Fix dragging active timeseries in dashboard.
+
 - Fix IE svg not scaling well, set height in css.
 
 - Fix IE dashoard dragging, include datalist polyfill and get attribute instead of element.dataset.
+
 - Fix respect to layer minZoom and maxZoom.
 
 - Login button visible if not authenticated.
@@ -16,6 +19,7 @@ Release 3.0.10 (2016-4-22)
 ---------------------
 
 - Download rain data directly from 5minute raster endpoint.
+
 - Fix mixing timeseries of nested and parent assets.
 
 - Cancel consecutive calls for timeseries.